+++ webpack.config.js
@@ -20,7 +20,6 @@ const extensionConfig = {
     libraryTarget: 'commonjs2'
   },
   externals: {
-    azdata: 'azdata',
     vscode: 'commonjs vscode' // the vscode-module is created on-the-fly and must be excluded. Add other modules that cannot be webpack'ed, 📖 -> https://webpack.js.org/configuration/externals/
     // modules added here also need to be added in the .vscodeignore file
   },